fix(reactive-store): optional store reset on dismount
Do not destroy store on component dismount; as this kills future uses of the same hook.

* Force computed properties to be recomputed on re-mounting.
* Provide configuration `autoReset` option: to allow stores to reset to initial values when the associated hook/component is dismounted.

The `autoReset` [default value is ‘false’] is useful for situations such as routing;  where the reactive store should be preserved (by default). In other scenarios the reactive store can be transient and its values will be reset to the original values.